@@ -3,32 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
-    "@heroicons/react": "^2.1.5",
-    "@zxing/library": "^0.21.3",
-    "chart.js": "^4.4.4",
-    "file-saver": "^2.0.5",
-    "jsqr": "^1.4.0",
-    "jszip": "^3.10.1",
-    "next": "14.2.13",
-    "parse": "^5.3.0",
-    "qrcode-base64": "^1.0.1",
-    "qrcode.react": "^4.0.1",
-    "react": "^18",
-    "react-chartjs-2": "^5.2.0",
-    "react-dom": "^18",
-    "react-icons": "^5.3.0",
-    "react-qr-reader": "^3.0.0-beta-1",
-    "react-webcam": "^7.2.0",
-    "uuid": "^10.0.0"
+    "react": "19.0.0-rc-66855b96-20241106",
+    "react-dom": "19.0.0-rc-66855b96-20241106",
+    "next": "15.0.3"
   },
   "devDependencies": {
+    "typescript": "^5",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "postcss": "^8",
-    "tailwindcss": "^3.4.1"
+    "tailwindcss": "^3.4.1",
+    "eslint": "^8",
+    "eslint-config-next": "15.0.3"
   }
 }
